폴더 정리 및 xrp receive daemon 구체화

+++ src/main/java/com/axia/dao/master/XrpAccountRepo.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface XrpAccountRepo extends JpaRepository<XrpAccount, Long> {
+public interface XrpAccountRepo extends JpaRepository<XrpAccount, Integer> {
     Optional<XrpAccount> findFirstByAddressAndDestination(@NonNull String address, @NonNull int destination);
 
 }
